requirements.txt
flask
flask_SQLAlchemy
python-dotenv
requests
Werkzeug
flask
gunicorn
flask_login